Changed structure of Camera to work with only Graphics components

+++ Engine/Models/Components/IGraphicsComponent.cs
@@ -11,8 +11,8 @@ namespace Engine.Models.Components
     public interface IGraphicsComponent : IGameComponent
     {
         public Vector2 Position { get; set; }
-        public double Width { get; set; }
-        public double Height { get; set; }
+        public float Width { get; set; }
+        public float Height { get; set; }
         public ImgNames CurrentImageName { get; set; }
         public List<ImgNames> ImageNames { get; set; }
     }
